+++ RELEASES.md
@@ -10,6 +10,7 @@
 * Add `Clone` and `Debug` impls for `FastTranslator`. (#25)
 * Add `FastTranslator::from_genetic_code` and tweak `FastTranslator::from_table` to take
   table reference instead of value. (#26)
+* Add `Seq::translated_*` methods. (#30)
 * Allow `Seq` to be compared to strings. (#28)
 * Add `Symbol::seq` helper. (#29)